cron/requirements.txt
httpx==0.25.1
python-dotenv==1.2.1